requirements.txt
colorama
pillow
opencv-python
